rename: 그룹별 페이지 카테고리 페이지로 이름 변경
